Add support for ipykernel 6
In ipykernel 6, the shell_streams argument for the Kernel is deprecated. When
a list of shell_streams is provided, only the first one is actually used in
ipykernel 6, i.e. the control stream is not setup properly.
The shell_stream and control_stream are individual argument for the Kernel in
ipykernel 6.

Compatibilty with ipykernel 5 is preserved, but might be dropped later.

Upstream change:
https://github.com/ipython/ipykernel/blob/7ed57cb4a03e719074c373029cf9f2e6dfd2c6de/ipykernel/kernelbase.py#L75

Should solve the following issue:
https://github.com/albertz/background-zmq-ipython/issues/6