Do not listen for remote connections by default
Introduce the 'allow_remote_connections' argument for the init_ipython_kernel
method make listing for remote connections from other machines optional.
The default is not allow remote connections to avoid security issues when
the feature is not required. On Windows that means that by default
the dialog asking the user to set a firewall rule is not shown when
initializing the kernel.

Should resolve the issue:
https://github.com/albertz/background-zmq-ipython/issues/10